+++ frontend/src/mocks/ingredients.json
@@ -1,17 +1,17 @@
 [
-  { "id": 1, "name": "Грибы","class":"filling--mushrooms" , "image": "filling/mushrooms.svg", "price": 33},
-  { "id": 2, "name": "Чеддер", "class":"filling--cheddar" , "image": "filling/cheddar.svg", "price": 42 },
-  { "id": 3, "name": "Салями", "class":"filling--salami" ,"image": "filling/salami.svg", "price": 42 },
-  { "id": 4, "name": "Ветчина", "class":"filling--ham"  ,"image": "filling/ham.svg", "price": 42 },
-  { "id": 5, "name": "Ананас", "class":"filling--ananas" , "image": "filling/ananas.svg", "price": 25 },
-  { "id": 6, "name": "Бекон", "class":"filling--bacon" , "image": "filling/bacon.svg", "price": 42 },
-  { "id": 7, "name": "Лук", "class":"filling--onion" , "image": "filling/onion.svg", "price": 21 },
-  { "id": 8, "name": "Чили", "class":"filling--chile" , "image": "filling/chile.svg", "price": 21 },
-  { "id": 9, "name": "Халапеньо", "class":"filling--jalapeno" , "image": "filling/jalapeno.svg", "price": 25 },
-  { "id": 10, "name": "Маслины", "class":"filling--olives" , "image": "filling/olives.svg", "price": 25 },
-  { "id": 11, "name": "Томаты", "class":"filling--tomatoes" , "image": "filling/tomatoes.svg", "price": 35 },
-  { "id": 12, "name": "Лосось", "class":"filling--salmon" , "image": "filling/salmon.svg", "price": 50 },
-  { "id": 13, "name": "Моцарелла", "class":"filling--mozzarella" , "image": "filling/mozzarella.svg", "price": 35 },
-  { "id": 14, "name": "Пармезан", "class":"filling--parmesan" , "image": "filling/parmesan.svg", "price": 35 },
-  { "id": 15, "name": "Блю чиз", "class":"filling--blue_cheese" , "image": "/img/filling/blue_cheese.svg", "price": 50 }
+  { "id": 1, "name": "Грибы","value":"mushrooms" , "image": "filling/mushrooms.svg", "price": 33},
+  { "id": 2, "name": "Чеддер", "value":"cheddar" , "image": "filling/cheddar.svg", "price": 42 },
+  { "id": 3, "name": "Салями", "value":"salami" ,"image": "filling/salami.svg", "price": 42 },
+  { "id": 4, "name": "Ветчина", "value":"ham"  ,"image": "filling/ham.svg", "price": 42 },
+  { "id": 5, "name": "Ананас", "value":"ananas" , "image": "filling/ananas.svg", "price": 25 },
+  { "id": 6, "name": "Бекон", "value":"bacon" , "image": "filling/bacon.svg", "price": 42 },
+  { "id": 7, "name": "Лук", "value":"onion" , "image": "filling/onion.svg", "price": 21 },
+  { "id": 8, "name": "Чили", "value":"chile" , "image": "filling/chile.svg", "price": 21 },
+  { "id": 9, "name": "Халапеньо", "value":"jalapeno" , "image": "filling/jalapeno.svg", "price": 25 },
+  { "id": 10, "name": "Маслины", "value":"olives" , "image": "filling/olives.svg", "price": 25 },
+  { "id": 11, "name": "Томаты", "value":"tomatoes" , "image": "filling/tomatoes.svg", "price": 35 },
+  { "id": 12, "name": "Лосось", "value":"salmon" , "image": "filling/salmon.svg", "price": 50 },
+  { "id": 13, "name": "Моцарелла", "value":"mozzarella" , "image": "filling/mozzarella.svg", "price": 35 },
+  { "id": 14, "name": "Пармезан", "value":"parmesan" , "image": "filling/parmesan.svg", "price": 35 },
+  { "id": 15, "name": "Блю чиз", "value":"blue_cheese" , "image": "/img/filling/blue_cheese.svg", "price": 50 }
 ]
