+++ frontend/src/mocks/sauces.json
@@ -2,11 +2,13 @@
   {
     "id": 1,
     "name": "Томатный",
-    "price": 50
+    "price": 50,
+    "value": "tomato"
   },
   {
     "id": 2,
     "name": "Сливочный",
-    "price": 50
+    "price": 50,
+    "value": "creamy"
   }
 ]
